# simplify push logic