--- a/go.mod
+++ b/go.mod
@@ -4,7 +4,9 @@ go 1.11
 
 require (
 	github.com/christianrpetrin/queue-tests v0.0.0-20181113195552-8536b5b8f660
+	github.com/davecgh/go-spew v1.1.1
 	github.com/gammazero/deque v0.0.0-20180920172122-f6adf94963e4
+	github.com/kr/pretty v0.1.0
 	github.com/phf/go-queue v0.0.0-20170504031614-9abe38d0371d
 	gopkg.in/karalabe/cookiejar.v2 v2.0.0-20150724131613-8dcd6a7f4951
 )