--- a/unit_test.go
+++ b/unit_test.go
@@ -21,7 +21,10 @@
 package deque
 
 import (
+	"fmt"
 	"testing"
+
+	"github.com/davecgh/go-spew/spew"
 )
 
 const (
@@ -31,33 +34,24 @@ const (
 
 func TestNewShouldReturnInitiazedInstanceOfDeque(t *testing.T) {
 	d := New()
+	assertInvariants(t, d, nil)
+}
 
-	if d == nil {
-		t.Error("Expected: new instance of queue; Got: nil")
+func TestInvariants(t *testing.T) {
+	d := New()
+	assertInvariants(t, d, nil)
+	for i := 0; i < maxInternalSliceSize*10; i++ {
+		d.PushBack(i)
+		assertInvariants(t, d, func(i int) interface{} {
+			return i
+		})
 	}
-	if d.Len() != 0 {
-		t.Errorf("Expected: %d; Got: %d", 0, d.Len())
-	}
-	if _, ok := d.Back(); ok {
-		t.Error("Expected: false; Got: true")
-	}
-	if _, ok := d.Front(); ok {
-		t.Error("Expected: false; Got: true")
-	}
-	if d.head != nil {
-		t.Error("Expected: d.head == nil; Got: d.head != nil")
-	}
-	if d.tail != nil {
-		t.Error("Expected: d.tail == nil; Got: d.tail != nil")
-	}
-	if d.tp != 0 {
-		t.Errorf("Expected: %d; Got: %d", 0, d.tp)
-	}
-	if d.hp != 0 {
-		t.Errorf("Expected: %d; Got: %d", 0, d.hp)
-	}
-	if d.spareLinks != 0 {
-		t.Errorf("Expected: %d; Got: %d", 0, d.spareLinks)
+
+	for i := 0; i < maxInternalSliceSize*10; i++ {
+		d.PopBack()
+		assertInvariants(t, d, func(i int) interface{} {
+			return i
+		})
 	}
 }
 
@@ -760,28 +754,180 @@ func TestPopBackWithRefillShouldKeepMaxSpareLinks(t *testing.T) {
 
 // Checks the internal slices and its links.
 func checkLinks(t *testing.T, d *Deque, length, headSliceSize, tailSliceSize, spareLinks int, headNext, headPrevious, tailNext, tailPrevious *node) {
+	t.Helper()
 	if d.Len() != length {
-		t.Errorf("Expected: %d; Got: %d", length, d.Len())
+		t.Errorf("unexpected length; Expected: %d; Got: %d", length, d.Len())
 	}
 	if len(d.head.v) != headSliceSize {
-		t.Errorf("Expected: %d; Got: %d", headSliceSize, len(d.head.v))
+		t.Errorf("unexpected head size; Expected: %d; Got: %d", headSliceSize, len(d.head.v))
 	}
 	if d.head.n != headNext {
-		t.Error("Expected: d.head.n == headNext; Got: d.head.n != headNext")
+		t.Error("unexpected head node; Expected: d.head.n == headNext; Got: d.head.n != headNext")
 	}
 	if d.head.p != headPrevious {
-		t.Error("Expected: d.head.p == headPrevious; Got: d.head.p != headPrevious")
+		t.Error("unexpected head; Expected: d.head.p == headPrevious; Got: d.head.p != headPrevious")
 	}
 	if d.tail.n != tailNext {
-		t.Error("Expected: d.tail.n == tailNext; Got: d.tail.n != tailNext")
+		t.Error("unexpected tailNext; Expected: d.tail.n == tailNext; Got: d.tail.n != tailNext")
 	}
 	if d.tail.p != tailPrevious {
-		t.Error("Expected: d.tail.p == tailPrevious; Got: d.tail.p != tailPrevious")
+		t.Error("unexpected tailPrevious; Expected: d.tail.p == tailPrevious; Got: d.tail.p != tailPrevious")
 	}
 	if len(d.tail.v) != tailSliceSize {
-		t.Errorf("Expected: %d; Got: %d", tailSliceSize, len(d.tail.v))
+		t.Errorf("unexpected tail size; Expected: %d; Got: %d", tailSliceSize, len(d.tail.v))
 	}
 	if d.spareLinks != spareLinks {
-		t.Errorf("Expected: %d; Got: %d", spareLinks, d.spareLinks)
+		t.Errorf("unexpected spare link count; Expected: %d; Got: %d", spareLinks, d.spareLinks)
+	}
+	if t.Failed() {
+		t.FailNow()
+	}
+}
+
+// assertInvariants checks all the invariant conditions in d that we can think of.
+// If val is non-nil it is used to find the expected value for an item at index
+// i measured from the head of the queue.
+func assertInvariants(t *testing.T, d *Deque, val func(i int) interface{}) {
+	t.Helper()
+	fail := func(what string, got, want interface{}) {
+		t.Log(spew.Sdump(d))
+		t.Errorf("invariant fail: %s; got %v want %v", what, got, want)
+	}
+	if d == nil {
+		fail("non-nil Deque", d, "non-nil")
+	}
+	if d.head == nil {
+		// Zero value.
+		if d.tail != nil {
+			fail("nil tail when zero", d.tail, nil)
+		}
+		if d.len != 0 {
+			fail("zero length when zero", d.len, 0)
+		}
+		if d.hp != 0 {
+			fail("zero hp when zero", d.hp, 0)
+		}
+		if d.tp != 0 {
+			fail("zero tp when zero", d.tp, 0)
+		}
+		if d.spareLinks != 0 {
+			fail("no spare links when zero", d.spareLinks, 0)
+		}
+		return
+	}
+	if d.head == d.tail {
+		if d.len == 0 {
+			if d.tp != -1 {
+				fail("tail index == -1 when empty", d.tp, -1)
+			}
+			if d.hp != 0 {
+				fail("head index == 0 when empty", d.hp, 0)
+			}
+		} else {
+			if d.hp > d.tp {
+				fail("head index exceeds tail index", d.hp, d.tp)
+			}
+		}
+	}
+	spareLinkCount := 0
+	inQueue := true
+	elemCount := 0
+	smallNodeCount := 0
+	index := 0
+	walkLinks(t, d, func(n *node) {
+		if len(n.v) < maxInternalSliceSize {
+			smallNodeCount++
+			if len(n.v) > maxFirstSliceSize {
+				fail("first node within bounds", len(n.v), maxFirstSliceSize)
+			}
+		}
+		if len(n.v) > maxInternalSliceSize {
+			fail("slice too big", len(n.v), maxInternalSliceSize)
+		}
+		for i, v := range n.v {
+			failElem := func(what string, got, want interface{}) {
+				fail(fmt.Sprintf("at elem %d, node %p, %s", i, n, what), got, want)
+				t.FailNow()
+			}
+			if !inQueue {
+				if v != nil {
+					failElem("all values outside queue nil", v, nil)
+				}
+				continue
+			}
+			elemInQueue := inQueue
+			switch {
+			case n == d.head && n == d.tail:
+				elemInQueue = i >= d.hp && i <= d.tp
+			case n == d.head:
+				elemInQueue = i >= d.hp
+			case n == d.tail:
+				elemInQueue = i <= d.tp
+			}
+			if elemInQueue {
+				if v == nil {
+					failElem("all values inside queue non-nil", v, "non-nil")
+				}
+			} else {
+				if v != nil {
+					failElem("all values outside queue nil", v, nil)
+				}
+			}
+			if v != nil {
+				if val != nil {
+					want := val(index)
+					if want != v {
+						failElem(fmt.Sprintf("element %d has expected value", index), v, want)
+					}
+				}
+				elemCount++
+				index++
+			}
+		}
+		if !inQueue {
+			spareLinkCount++
+		}
+		if n == d.tail {
+			inQueue = false
+		}
+	})
+	if inQueue {
+		// We never encountered the tail pointer.
+		t.Errorf("tail does not point to element in list")
+	}
+	if spareLinkCount > maxSpareLinks {
+		fail("spare link count <= maxSpareLinks", spareLinkCount, maxSpareLinks)
+	}
+	if elemCount != d.len {
+		fail("element count == d.len", elemCount, d.len)
+	}
+	if smallNodeCount > 1 {
+		fail("only one first node", smallNodeCount, 1)
+	}
+	if t.Failed() {
+		t.FailNow()
+	}
+}
+
+// walkLinks calls f for each node in the linked list.
+// It also checks link invariants:
+func walkLinks(t *testing.T, d *Deque, f func(n *node)) {
+	t.Helper()
+	fail := func(what string, got, want interface{}) {
+		t.Errorf("link invariant %s fail; got %v want %v", what, got, want)
+	}
+	n := d.head
+	for {
+		if n.n.p != n {
+			fail("node.n.p == node", n.n.p, n)
+		}
+		if n.p.n != n {
+			fail("node.p.n == node", n.p.n, n)
+		}
+		f(n)
+		n = n.n
+		if n == d.head {
+			break
+		}
 	}
 }