The lastTailPosition field is unnecessary, as it's only used for the first
block, and we can always tell if we're on the first block by checking the
block length.

Initially, this code changed some invariants which caused the unit tests
to fail. Although this was an issue with the code, the reason for the unit
test failure was obscure, and changing the invariants would require
changing a lot of fairly obscure unit tests. As a possible alternative approach,
I've added `assertInvariants`, a function which checks all known
invariants and could be called at any point. This also provides some
reasonable documentation of what the invariants actually are (it
would be nice if they were also documented better in the code).

Then the code could be refactored more freely while only changing
a single invariant checker function.

As an aside, it's not clear why we have different max size behaviour for
the first slice versus the rest. Why not let the first slice expand to 256
instead of 16? That would simplify the code a little more.

Performance seems generally better or around the same:

	benchmark                                     old ns/op     new ns/op     delta
	BenchmarkFillDequeQueue/0-4                   42.6          41.0          -3.76%
	BenchmarkFillDequeQueue/1-4                   178           151           -15.17%
	BenchmarkFillDequeQueue/10-4                  712           687           -3.51%
	BenchmarkFillDequeQueue/100-4                 5367          5351          -0.30%
	BenchmarkFillDequeQueue/1000-4                44308         44448         +0.32%
	BenchmarkFillDequeQueue/10000-4               436426        468247        +7.29%
	BenchmarkFillDequeQueue/100000-4              4701266       5020198       +6.78%
	BenchmarkFillDequeQueue/1000000-4             52650621      57124863      +8.50%
	BenchmarkFillDequeStack/0-4                   43.0          40.5          -5.81%
	BenchmarkFillDequeStack/1-4                   163           160           -1.84%
	BenchmarkFillDequeStack/10-4                  704           726           +3.12%
	BenchmarkFillDequeStack/100-4                 6206          5500          -11.38%
	BenchmarkFillDequeStack/1000-4                82310         43529         -47.12%
	BenchmarkFillDequeStack/10000-4               558683        430710        -22.91%
	BenchmarkFillDequeStack/100000-4              5157938       4695966       -8.96%
	BenchmarkFillDequeStack/1000000-4             57210349      55608442      -2.80%
	BenchmarkMicroserviceDequeQueue/1-4           617           563           -8.75%
	BenchmarkMicroserviceDequeQueue/10-4          4345          4070          -6.33%
	BenchmarkMicroserviceDequeQueue/100-4         30772         28047         -8.86%
	BenchmarkMicroserviceDequeQueue/1000-4        263054        272522        +3.60%
	BenchmarkMicroserviceDequeQueue/10000-4       2794353       2834015       +1.42%
	BenchmarkMicroserviceDequeQueue/100000-4      30656758      30792600      +0.44%
	BenchmarkMicroserviceDequeQueue/1000000-4     310948195     322120527     +3.59%
	BenchmarkMicroserviceDequeStack/1-4           528           457           -13.45%
	BenchmarkMicroserviceDequeStack/10-4          3405          3060          -10.13%
	BenchmarkMicroserviceDequeStack/100-4         30136         27090         -10.11%
	BenchmarkMicroserviceDequeStack/1000-4        349095        260084        -25.50%
	BenchmarkMicroserviceDequeStack/10000-4       4842273       2779491       -42.60%
	BenchmarkMicroserviceDequeStack/100000-4      30376486      30133643      -0.80%
	BenchmarkMicroserviceDequeStack/1000000-4     403197125     314148386     -22.09%
	BenchmarkRefillFullDequeQueue/1-4             5306          4137          -22.03%
	BenchmarkRefillFullDequeQueue/10-4            40988         41801         +1.98%
	BenchmarkRefillFullDequeQueue/100-4           556997        422356        -24.17%
	BenchmarkRefillFullDequeQueue/1000-4          5747318       4039626       -29.71%
	BenchmarkRefillFullDequeQueue/10000-4         50574502      49336337      -2.45%
	BenchmarkRefillFullDequeQueue/100000-4        480063868     496396217     +3.40%
	BenchmarkRefillFullDequeStack/1-4             4972          4494          -9.61%
	BenchmarkRefillFullDequeStack/10-4            40274         43639         +8.36%
	BenchmarkRefillFullDequeStack/100-4           394611        396665        +0.52%
	BenchmarkRefillFullDequeStack/1000-4          3978534       4017593       +0.98%
	BenchmarkRefillFullDequeStack/10000-4         45048073      44497156      -1.22%
	BenchmarkRefillFullDequeStack/100000-4        463631405     473849367     +2.20%
	BenchmarkRefillDequeQueue/1-4                 4259          4487          +5.35%
	BenchmarkRefillDequeQueue/10-4                40429         46295         +14.51%
	BenchmarkRefillDequeQueue/100-4               395224        430059        +8.81%
	BenchmarkRefillDequeQueue/1000-4              3951623       4097020       +3.68%
	BenchmarkRefillDequeQueue/10000-4             43354326      47419649      +9.38%
	BenchmarkRefillDequeQueue/100000-4            474838615     507786535     +6.94%
	BenchmarkRefillDequeStack/1-4                 4187          4195          +0.19%
	BenchmarkRefillDequeStack/10-4                40136         43958         +9.52%
	BenchmarkRefillDequeStack/100-4               388252        414602        +6.79%
	BenchmarkRefillDequeStack/1000-4              3813335       4212829       +10.48%
	BenchmarkRefillDequeStack/10000-4             42989920      47718678      +11.00%
	BenchmarkRefillDequeStack/100000-4            470917806     507785491     +7.83%
	BenchmarkSlowDecreaseDequeQueue/1-4           40.4          42.4          +4.95%
	BenchmarkSlowDecreaseDequeQueue/10-4          407           454           +11.55%
	BenchmarkSlowDecreaseDequeQueue/100-4         3951          4747          +20.15%
	BenchmarkSlowDecreaseDequeQueue/1000-4        40705         41896         +2.93%
	BenchmarkSlowDecreaseDequeQueue/10000-4       403659        430482        +6.64%
	BenchmarkSlowDecreaseDequeQueue/100000-4      3957679       4371165       +10.45%
	BenchmarkSlowDecreaseDequeQueue/1000000-4     38944592      42589633      +9.36%
	BenchmarkSlowDecreaseDequeStack/1-4           47.9          42.6          -11.06%
	BenchmarkSlowDecreaseDequeStack/10-4          512           431           -15.82%
	BenchmarkSlowDecreaseDequeStack/100-4         5033          4232          -15.91%
	BenchmarkSlowDecreaseDequeStack/1000-4        38407         41127         +7.08%
	BenchmarkSlowDecreaseDequeStack/10000-4       394338        424500        +7.65%
	BenchmarkSlowDecreaseDequeStack/100000-4      4137764       4159700       +0.53%
	BenchmarkSlowDecreaseDequeStack/1000000-4     39012681      41832207      +7.23%
	BenchmarkSlowIncreaseDequeQueue/1-4           340           283           -16.76%
	BenchmarkSlowIncreaseDequeQueue/10-4          2280          2309          +1.27%
	BenchmarkSlowIncreaseDequeQueue/100-4         9023          9635          +6.78%
	BenchmarkSlowIncreaseDequeQueue/1000-4        84790         87842         +3.60%
	BenchmarkSlowIncreaseDequeQueue/10000-4       1054093       888949        -15.67%
	BenchmarkSlowIncreaseDequeQueue/100000-4      11396544      10057962      -11.75%
	BenchmarkSlowIncreaseDequeQueue/1000000-4     115985396     106593814     -8.10%
	BenchmarkSlowIncreaseDequeStack/1-4           336           286           -14.88%
	BenchmarkSlowIncreaseDequeStack/10-4          1105          1138          +2.99%
	BenchmarkSlowIncreaseDequeStack/100-4         11422         9511          -16.73%
	BenchmarkSlowIncreaseDequeStack/1000-4        90291         81974         -9.21%
	BenchmarkSlowIncreaseDequeStack/10000-4       973651        839611        -13.77%
	BenchmarkSlowIncreaseDequeStack/100000-4      11435655      9859417       -13.78%
	BenchmarkSlowIncreaseDequeStack/1000000-4     101713693     108145656     +6.32%
	BenchmarkStableDequeQueue/1-4                 43.5          43.0          -1.15%
	BenchmarkStableDequeQueue/10-4                417           420           +0.72%
	BenchmarkStableDequeQueue/100-4               3981          4212          +5.80%
	BenchmarkStableDequeQueue/1000-4              39733         40513         +1.96%
	BenchmarkStableDequeQueue/10000-4             408448        404425        -0.98%
	BenchmarkStableDequeQueue/100000-4            3959589       4175790       +5.46%
	BenchmarkStableDequeQueue/1000000-4           39721862      41249243      +3.85%
	BenchmarkStableDequeStack/1-4                 44.2          44.1          -0.23%
	BenchmarkStableDequeStack/10-4                451           434           -3.77%
	BenchmarkStableDequeStack/100-4               4413          4417          +0.09%
	BenchmarkStableDequeStack/1000-4              41397         42712         +3.18%
	BenchmarkStableDequeStack/10000-4             420809        436227        +3.66%
	BenchmarkStableDequeStack/100000-4            4339661       4312148       -0.63%
	BenchmarkStableDequeStack/1000000-4           42356775      43244007      +2.09%