@christianrpetrin 2018-11-29
> As an aside, it's not clear why we have different max size behaviour for
> the first slice versus the rest. Why not let the first slice expand to 256
> instead of 16? That would simplify the code a little more.

I agree the implementation would be simpler. However, I ran a bunch of tests and the results are generally worse.

When it comes to the benchmark tests, well, there's just too many of them! It's hard to look at very single one of them comparing deque with every other queue. In my experience, with almost every change aimed at improving performance, you will see some gains in some tests and loses in others. It's sometimes hard then to figure out if the net impact was positive or not. So I generally favor comparing the results of the microservice test vs the slice implementation vs the cookiejar deque. If the results for these tests look good, I generally just glance over all the other tests to make sure there's nothing looking too bad or abnormal. The slice one is very fast for small data sets, but slow for large ones; cookiejar is the opposite. Our goal should always be to improve deque performance and efficiency so deque compares well especially when comparing to these two. We should aim to be faster or very competitive with slice with small data sets (<=100) and faster or very competitive with cookiejar for mid to large data sets (>=1000). Any idea why deque is generally slower than cookiejar for 1000+ items? It's because at 4 spare slices * 256 each one, deque caches at most 1024 usable slice positions. Past that, especially in refill situations, deque has to allocate more slices, while cookiejar deque, with its monster 4096 sized slices can cache a lot more before it has to allocate more slices. Increasing either spare slices or max internal slice size helps with large data sets, but is generally worse with small ones.